// ui-service/src/app/components/match/model/match-column-model.ts
export enum MatchColumn {
  ID = 'id',
  SOURCE_SYSTEM_NAME = 'sourceSystemName',
  SOURCE_SYSTEM_ID = 'sourceSystemId',
  HOME_TEAM_NAME = 'homeTeamName',
  AWAY_TEAM_NAME = 'awayTeamName',
  START_TIME = 'startTime',
  RESULT = 'result',
  RESULT_TYPE_LBL = 'resultTypeLbl',
  MARKED_AS_FINISHED_TIME = 'markedAsFinishedTime'

}

export enum MatchOddColumn {
  BOOKMAKER_NAME = 'bookmakerName',
  UPDATED_ON = 'updatedOn',
  ODD1 = 'odd1',
  ODD2 = 'odd2',
  ODDX = 'oddX',
  ODD1X = 'odd1X',
  ODD12 = 'odd12',
  ODDX2 = 'oddX2',
  ODDBTSY = 'oddBTSY',
  ODDBTSN = 'oddBTSN',
  ODDO05 = 'oddO05',
  ODDO15 = 'oddO15',
  ODDO25 = 'oddO25',
  ODDO35 = 'oddO35',
  ODDO45 = 'oddO45',
  ODDO55 = 'oddO55',
  ODDO65 = 'oddO65',
  ODDU05 = 'oddU05',
  ODDU15 = 'oddU15',
  ODDU25 = 'oddU25',
  ODDU35 = 'oddU35',
  ODDU45 = 'oddU45',
  ODDU55 = 'oddU55',
  ODDU65 = 'oddU65',
}